Split observer detail components into per-component folders
ObserverDetails and ObserversForEventType were single-file components
that combined a composition root with multiple rendering concerns. Per
the React component conventions, multi-concern components belong in a
folder of their own with sub-components, co-located CSS, an index
barrel, and a README.

ObserverDetails now composes ObserverSummary and ObserverEventTypes,
ObserversForEventType composes ObserversForEventTypeTable. Inline style
blocks move into co-located CSS files using PrimeReact design tokens.
