+++ Source/Workbench/Features/EventStore/General/EventTypes/ObserversForEventType/ObserversForEventTypeTable.css
@@ -1,0 +1,3 @@
+.observers-for-event-type-table {
+    height: 100%;
+}
