+++ Source/Workbench/Features/EventStore/General/EventTypes/ObserversForEventType/index.ts
@@ -1,0 +1,5 @@
+// Copyright (c) Cratis. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+export * from './ObserversForEventType';
+export * from './ObserversForEventTypeTable';
