+++ Source/Workbench/Features/EventStore/Namespaces/Observers/ObserverDetails/ObserverEventTypes.css
@@ -1,0 +1,5 @@
+.observer-event-types {
+    flex: 1;
+    min-height: 0;
+    padding: 0 16px 16px 16px;
+}
