+++ Source/Workbench/Features/EventStore/Namespaces/Observers/ObserverDetails/index.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Cratis. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+export * from './ObserverDetails';
+export * from './ObserverSummary';
+export * from './ObserverEventTypes';
